+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @2023-ignite-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- Add Tooltip component
+
 ## 2.0.1
 
 ### Patch Changes