+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @2023-ignite-ui/tokens
 
+## 1.2.0
+
+### Minor Changes
+
+- Add Tooltip component
+
 ## 1.1.0
 
 ### Minor Changes